feat(js): moved all web socket connection code under ws folder
